move Input-related constants.ts to components/form/
